pyproject.toml
[project]
name = "fakturoid-connector"
version = "0.1.0"
description = "CLI tool, MCP server, and Discord notifier for Fakturoid.cz"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Continero Corp s.r.o.", email = "info@continero.com"},
]
keywords = ["fakturoid", "invoicing", "mcp", "cli", "discord"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Accounting",
]
dependencies = [
    "requests>=2.31",
    "click>=8.1",
    "python-dotenv>=1.0",
    "mcp[cli]>=1.0",
]

[project.urls]
Homepage = "https://github.com/Continero/fakturoid-connector"
Repository = "https://github.com/Continero/fakturoid-connector"
Issues = "https://github.com/Continero/fakturoid-connector/issues"

[project.scripts]
fakturoid = "fakturoid_connector.cli:cli"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
